Moved input components into component/input
