+++ core/src/com/brinkman/platformer/component/input/InputOperator.java
@@ -1,13 +1,14 @@
-package com.brinkman.platformer.component;
+package com.brinkman.platformer.component.input;
 
 import com.brinkman.platformer.GameWorld;
+import com.brinkman.platformer.component.Operator;
+import com.brinkman.platformer.component.RootComponent;
 import com.brinkman.platformer.component.physics.PhysicsComponent;
 import com.brinkman.platformer.component.render.AnimationType;
 import com.brinkman.platformer.component.render.RenderComponent;
 import com.brinkman.platformer.entity.Entity;
 import com.brinkman.platformer.input.ControllerProcessor;
 import com.brinkman.platformer.input.KeyboardProcessor;
-import com.brinkman.platformer.physics.Body;
 import com.brinkman.platformer.physics.ControlledBody;
 
 import java.util.Collection;
