+++ core/src/com/brinkman/platformer/entity/actor/Player.java
@@ -9,8 +9,8 @@ import com.badlogic.gdx.graphics.g2d.TextureRegion;
 import com.badlogic.gdx.math.Vector2;
 import com.badlogic.gdx.utils.Array;
 import com.badlogic.gdx.utils.Logger;
-import com.brinkman.platformer.component.InputComponent;
-import com.brinkman.platformer.component.PlayerInputComponent;
+import com.brinkman.platformer.component.input.InputComponent;
+import com.brinkman.platformer.component.input.PlayerInputComponent;
 import com.brinkman.platformer.component.physics.ControlledPhysicsComponent;
 import com.brinkman.platformer.component.physics.PhysicsComponent;
 import com.brinkman.platformer.component.render.AnimationRenderComponent;
